+++ Instance.h
@@ -24,10 +24,10 @@ public:
 
     string name;
     int n;
-    int* lengths;
+    double* lengths;
     double* halfLengths;
     int** demands;
-    int layoutLengh;
+    double layoutLengh;
 
 };
